refactor(mpsc): simplify unbounded channel coordination
